+++ private/buf/cmd/buf/command/generate/generate.go
@@ -538,7 +538,7 @@ func run(
 		flags.Config,
 		flags.Paths,
 		flags.ExcludePaths,
-		flags.Types,
+		append(flags.Types, flags.TypesDeprecated...),
 		flags.ExcludeTypes,
 	)
 	if err != nil {